+++ skrl/trainers/torch/base.py
@@ -6,7 +6,7 @@ import tqdm
 
 import torch
 
-from skrl import logger
+from skrl import config, logger
 from skrl.agents.torch import Agent
 from skrl.envs.wrappers.torch import Wrapper
 
@@ -74,6 +74,11 @@ class Trainer:
                 logger.info("Closing environment")
                 self.env.close()
                 logger.info("Environment closed")
+
+        # update trainer configuration to avoid duplicated info/data in distributed runs
+        if config.torch.is_distributed:
+            if config.torch.rank:
+                self.disable_progressbar = True
 
     def __str__(self) -> str:
         """Generate a string representation of the trainer